+++ src/api/resources/queries/char/allReqChars.ts
@@ -1,0 +1,57 @@
+import gql from "graphql-tag";
+
+export const ALL_REQ_CHARS = gql`
+  {
+    allRequestedChars {
+      id
+      author
+      _id
+      name
+      field_diameter
+      move_type
+      type
+      has_visible_level
+      char_pic
+      mob_range
+      is_mob_aggressive
+      monologs {
+        _id
+        content
+      }
+      dialogs {
+        _id
+        npc
+        player {
+          _id
+          dialog
+          action
+          condition
+          next
+        }
+      }
+      settings {
+        time_of_occurance {
+          min
+          max
+        }
+        resp_time {
+          min
+          max
+        }
+      }
+      statistics {
+        strength
+        speed
+        level
+        jink
+        inteligence
+        health
+        dexterity
+        defence
+        attack
+        attack_range
+        attack_speed
+      }
+    }
+  }
+`;