+++ src/api/resources/queries/map/allMapsBaseInfo.ts
@@ -1,0 +1,11 @@
+import gql from "graphql-tag";
+
+export const ALL_MAPS_BASE_INFO = gql`
+  {
+    allMaps {
+      id
+      _id
+      map_name
+    }
+  }
+`;